remove Push{Keys,Values}, rename Slice, add Pointer, remove heap dependency
The only reason the heap dependency existed was because On initialized
the heap before returning. The sliceheap.Heap can be more generally
useful if it does not require that dependency.

Removes Push* to accomodate that; they were nifty but only replaced a
simple loop.

Slice was misleading because it was only a temporary view, so the name
changed. Pointer exists now to provide a permanent view.